Bug#34011717 socket::is_open() is true after socket::close() fails
Issue
=====

socket::is_open() checks if a socket is open by checking if the socket
descriptor is assigned.

After socket::close() succeeds the socket-descriptor is reset.

In case socket::close() fails, the socket-descriptor's state is
currently unchanged.

Change
======

According to "man close" on Linux, a failure of "close()" leaves the
socket in an undefined state and that it shouldn't be retried. Instead
it should be logged and treated as closed.

- reset socket-descriptor at socket::close() even if close() fails.

Change-Id: I0c70af17317bf205504909d66403f6c388ab1a6e