WL#15130 Socket-level TLS patch #11: Use NdbSocket in mgmapi.cpp
Use NdbSocket in the management API, except for in the Event
Listener.

If TLS is enabled on an mgm api socket, the client is not allowed
to upgrade that mgm connection to a transporter. Add a new error
code 1014 NDB_MGM_CANNOT_CONVERT_TO_TRANSPORTER to the mgm api
for this case.

Change-Id: I6454ce16379b1cc9919c5a9094ee9c550d14ae84