Bug#33616174: Querying data dictionary table KEY_COLUMN_USAGE is slow with hypergraph enabled [improve trace, noclose]
This patch improves the optimizer trace for queries with LATERAL to
make it easier to debug the query in the bug report.

Currently the "Enumerating subplans" part of the trace only prints
headers for subjoins if it is possible to use a hash join. If only
nested loop joins are considered, no such header is printed, which
makes it harder to read the trace. Take for example:

CREATE TABLE t1 (a INT, b INT);
CREATE TABLE t2 (a INT, b INT);
SELECT 1 FROM t1, LATERAL (SELECT SUM(b) FROM t2 WHERE t1.a = t2.a) AS derived;

The trace previously looked like this:

,----
| Enumerating subplans:
|
| Found node derived [rows=2]
|  - {STREAM, cost=0.4, init_cost=0.0, rows=1.0, parm={t1}} is first alternative, keeping
|  - current access paths for {derived}: {STREAM, cost=0.4, init_cost=0.0, rows=1.0, parm={t1}})
|
| Found node t1 [rows=1]
|  - {TABLE_SCAN, cost=0.2, init_cost=0.0, rows=1.0} is first alternative, keeping
|  - current access paths for {t1}: {TABLE_SCAN, cost=0.2, init_cost=0.0, rows=1.0})
|  - {NESTED_LOOP_JOIN, cost=0.7, init_cost=0.0, rows=1.0} is first alternative, keeping
|  - current access paths for {t1,derived}: {NESTED_LOOP_JOIN, cost=0.7, init_cost=0.0, rows=1.0})
`----

Notice that the NESTED_LOOP_JOIN path comes under the "Found node t1"
header, which is confusing.

This patch ensures the header is printed also when there is no
possible hash join, and the output now looks like this:

,----
| Enumerating subplans:
|
| Found node derived [rows=2]
|  - {STREAM, cost=0.4, init_cost=0.0, rows=1.0, parm={t1}} is first alternative, keeping
|  - current access paths for {derived}: {STREAM, cost=0.4, init_cost=0.0, rows=1.0, parm={t1}})
|
| Found node t1 [rows=1]
|  - {TABLE_SCAN, cost=0.2, init_cost=0.0, rows=1.0} is first alternative, keeping
|  - current access paths for {t1}: {TABLE_SCAN, cost=0.2, init_cost=0.0, rows=1.0})
|
| Found sets {t1} and {derived}, connected by condition (none)
|  - {NESTED_LOOP_JOIN, cost=0.7, init_cost=0.0, rows=1.0} is first alternative, keeping
|  - current access paths for {t1,derived}: {NESTED_LOOP_JOIN, cost=0.7, init_cost=0.0, rows=1.0})
`----

Change-Id: I1575b77ba8d71c26c76eeca602b58fb2b5fa0226