Bug#34009876: When table_open_cache_instances = 1, the result of               CONNECTION_ID() called on the trigger is wrong.
One might think that the CONNECTION_ID() function could be a constant
function, since it returns the session id that remains constant during
the lifetime of a session. However, when placed inside a trigger, which
is attached to a TABLE instance that may be reused by multiple sessions,
the function must pick up the current value of the session id.

Fixed by making the function const for execution and returning the
actual session id when the function is evaluated.

The test for this is placed in a new test file that executes with
table_open_cache_instances set to 1, in order to reuse TABLE entries
in subsequent sessions.

Change-Id: Ic820b553ea6a97091a6ff6748313a9c903cb4b9b