Bug #31750840 	ADAPTIVE HASH INDEX(AHI) BUILDING CAUSING CONTENTION ON BTR_SEARCH_LATCHES
The latches for searching in AHI were changed to non-blocking - we don't want to wait for a index which the only purpose is to speed up queries.
The latches for more update operations in AHI were changed to non-blocking:
- we don't have to insert new page to the index,
- we don't have to try to update a reference to a failed search,
we would do such operations next time we encounter them.

RB#27271

Change-Id: I68e4c2ec4b7fa3beebb0cea792f084f735fcd627