Bug#34025532 Check of pending signals by job scheduler is flaky
Patch change from using 'm_current_write_buffer_len' to calculate
the available 'number of words', to instead use the read_/write_index
pointer (Which are the same as we use to execute_signal() from the job buffers)

Intention of this sub-patch is to stop using the m_current_write_buffer_len,
which was never designed to be read by the consumer side.
Note that the added require from previous patch may still be hit.

Note that in case the producer side has started writing into a
new job_buffer page, beyond the job buffer page which the consumer is
currently executing from, we do NOT read the amount of 'words' filled
into this buffer. That would must likely have resulted in a cache line
miss, which is something we really try to avoid in this part of the code.
Instead we just take a half filled page as an estimate - likely any errors
made here will even out over time as the num_words estimate will be based
on multiple samples.

Change-Id: I3018200f021d98ea753c3b01101567f05503330a