Bug#33946721: Hypergraph orders and-terms inefficiently
Poor performance was seen in a query because table filters were
evaluated in a suboptimal way. The table filter was on this form:

  varchar_col LIKE '%foo' AND int_col = 15

Because the condition int_col = 15 is both more selective and cheaper
to evaluate than varchar_col LIKE '%foo', it is more beneficial to
evaluate int_col = 15 first, as that prevents most of the evaluations
of the more costly condition.

We don't have much information about the cost of evaluating a
predicate. We know if they contain subqueries or stored procedures,
which can potentially be very expensive, and the hypergraph optimizer
already moves such predicates last in the evaluation order.

We do have information about the selectivity of the predicates,
though. This patch improves the sorting of the WHERE predicates to
also take the selectivity into account, so that the more selective
predicates are evaluated before the less selective ones.

Change-Id: Icfae894c90848ad506018de147a85854deb20918