Support of POWER function in HEATWAVE
This patch adds the support of POWER function in HEATWAVE
Varlen data types are not supported.

DSB type input are casted to DOUBLE and input to POWER

Return type of POW is DOUBLE type and hence sometimes last decimal
mismatch is expected.
For Ex:
-6.2 is represented in HeatWave as -6.200000000000001

This is a known limitation of HeatWave documented.

Change-Id: I9ec909ccc2d4986edfabd3756e1462ac8715a6f3