Bug#34032102 Do not allocate job_buffers to links never used
Define an 'empty_job_buffer' object which is a job_buffer page of 'len==0'.

Init the job_buffers to refer either empty_job_buffer as a sentinel,
or to a 'nullptr' on the writer side. Set the write_buffer_len to
indicate that the write_buffer is 'full', such that nothing will
ever be written to the null-buffer.

Remove code for over-allocation of job_buffer pages to acount
for init of communication links never used.

Change-Id: Ibd765afd6b362105e604a407fe0ff6d0c69552de