Bug #33830934 Race between ut_lock_free_list_node<T>::grow and ut_lock_free_hash_t::get_tuple
This patch fixes three problems with ut_lock_free_hash_t.

1. The ut_lock_free_hash_t stores key-value pairs in buffers which it
allocates when previous buffer is full, and deallocates once all data is
moved from old buffer to one of the newer and there are no more
references to the old buffer.
Each buffer is represented as an instance of ut_lock_free_list_node_t
which has a field of type ut_lock_free_cnt_t which counts the number of
references.
For scalability there are 256 individual sub-counters, and a thread
increments one when entering a "critical section", and decrements one
when leaving it, and the thread doing reclamation checks if the sum is
zero.
Unfortunately, if a thread does inc() on a counter i but dec() on
counter i+k, then it may happen that the thread doing the summation in a
loop will see the decrement but not the increment and conclude the sum
is smaller than it should be. In particular it can appear that it is 0
while in reality the array is still in use. This crashes each time on
Apple M1 machine, because there we use RNG instead of CPU ID to
determine `i`, but in principle it could happen even with CPU IDs if the
thread switches from one CPU to another during "critical section".

The fix for this problem is to remember which of the sub-counters was
incremented and to decrement the same one. To make a mistake impossible,
this patch introduces a helper handle_t RAII-like class, which remembers
which of the counters should be released by it's destructor. An instance
of this class represents an "open handle" to the table, and can be moved
around to keep it open, but not copied.

2. I also found that the code has a race condition in ref-counting and
reclaiming a memory, as it tries to enforce "Store-Load barrier", a
thing which is notoriously hard to express/achieve. One thread wants to
Store a flag saying that array is scheduled for reclamation and then
Load the ref-counter, and another thread wants to Store the incremented
ref-counter and then load the pointer to data. I could not reproduce a
crash due to that, but I think the code is flawed anyway.

The fix here is to use seq_cst variant of atomic operations for
`arr->m_pending_free.store(true)` and for loads and increments of the
counters, and for m_pending_free.load().
This way, there must be some single total order on these operations,
which must be consistent with happens-before, which permits only two
possibilities:
i) m_pending_free.store(true) is ordered before m_pending_free.load(),
but then the loaded value must be `true` and the reader will refrain
from access.
ii) m_pending_free.store(true) is ordered after m_pending_free.load(),
but then counter.fetch_add(1) is ordered before counter.load() so the
non-zero value of the counter should prevent freeing the buffer.

3. There's insufficient memory ordering enforced between
a) ut_lock_free_list_node<T>::grow() operation which allocates memory,
constructs an object in it, and publishes a pointer to an object via
m_next atomic field, and
b) ut_lock_free_hash_t::get_tuple() which loads the m_next field, and
then tries to access the object pointed by it

In particular it seems to have a "classic" bug of publishing pointer to
newly initialized memory without enforcing correct ordering, so that
another thread can load the pointer to it, but still see the object as
uninitialized.

Take a look at `grow(...)`:

    next_t new_arr = ut::new_withkey<ut_lock_free_list_node_t<T>>(
        ut::make_psi_memory_key(mem_key_ut_lock_free_hash_t), new_size);

    /* Publish the allocated entry. If somebody did this in the
    meantime then just discard the allocated entry and do
    nothing. */
    next_t expected = nullptr;
    if (!m_next.compare_exchange_strong(expected, new_arr,
                                        std::memory_order_relaxed)) {

note the use of `std::memory_order_relaxed`: what it means is that we
*don't* want to establish synchronizes-with (and thus: happens-before)
relation with the thread which will read the `m_next`'s value stored by
us. Now take a look at the read side of things in `get_tuple(..)`:

   arr_node_t *next = (*arr)->m_next.load(std::memory_order_relaxed);
   (*arr)->end_access();
   if (next == nullptr) {
      return (nullptr);
   }
   *arr = next
 [....in the next loop iteration]:
   while (!(*arr)->begin_access()) {

here we load the pointer from `m_next` (again: resigning from
establishing happens-before) check it's not null (so far so good), and
try to call a method on the object pointed by it (apparently assuming,
that we already can see the results of the constructor of this object,
as opposed to some garbage or all-zeros!).

The fix for this bug is to use seq_cst variant of atomic operations, for
accesses of m_next, so that they properly establish happens-before.
Perhaps it would be sufficient to use release and acquire, but on ARM64
there's no additional cost for using seq_cst here, and the operation is
rare anyway.

4. The numa-specific code in ut0lock_free_hash.h was disabled by
quickfix to Bug#27792853 which simply setting the local member
`m_numa_available = false;` in the counter class.
The problem with this approach is that the number of counters was now
always set to 256, but the compiler did not attempt to optimize it as a
constant, because it wasn't literally a constant - it's still a member
field m_cnt_size, which just happened to be 256 always. This meant that
the `i%m_cnt_size` operation couldn't be optimized to simply extracting
the lowest significant byte of `i`.
Also, the code was simply allocating the atomic counters without any
padding, so it could happen that they got allocated close to each other.
In performance tests on tetra02 which is a NUMA machine with 90+ cpus,
it turned out that while the code for m_numa_available=true is indeed
very fast, it's wasting a lot of memory, because it uses
numa_alloc_onnode() which always allocates at least 4096 bytes, which
amounts to 1MB of memory just to store all 256 counters.
Also, the tests shown that it is not as important to allocate counters
to proper numa nodes - same performance gain could be obtained by simply
putting each counter in a different cache line.

The fix was to remove the unused code and to use cache-aligned atomics,
so that the code is 2x faster than the version with
m_numa_available=false, and as fast as old m_numa_available=true, but
doesn't waste so much space, nor time for modulo operations.

5. The m_base was a pointer to the actual array storing the key-value
pairs, and it was managed manually.

It was simply replaced with ut::unique_ptr<T[]> to simplify code.

6. On platforms such as Apple M1 where there is no easy way to find the
id of the CPU on which we are executing, we were using ut_rnd_gen_ulint
to pick the index of the subcounter to increment. This meant that the
same thread even if long running on the same CPU would try to modify
many different memory locations leading to contention.

The fix for this is to use this_thread::id() to compute the index,
which leads to observable performance improvement on Apple M1.

Change-Id: Ic5929cc949c51256f835e5a9502a6eb10a9d87ec