Bug#34038187 routertest_component_socket_close fails on pb2 (EL6)
The test changes something in the Router-perceived cluster topology
and then uses wait_for_transaction_count_increase(2) to make sure
it is be visible to the Router already. It assumes that each metadata
update consists of a single "strart transaction" statement which
is not the case for example when the Router updates last_check_in
field - then there are 2 "strart transaction" exectuted on a single
metadata server.

This patch increases the amount of the "start transaction" statements
the test waits to 3 to make sure the Router saw the cluster topology
changes.

Change-Id: I21c38a15dd7f242fa1861c47097dcdb923e0b080