Bug#33616174: Querying data dictionary table KEY_COLUMN_USAGE is slow with hypergraph enabled [lateral]
The hypergraph optimizer has some logic that narrows down the search
space when parameterized paths are involved. See
DisallowParameterizedJoinPath().

In order to work correctly, this code needs to know which tables are
directly reachable from the tables on each side of the current join.
The code that calculates the reachable tables does not take the
dependencies of lateral derived tables into account, and it may report
that a lateral derived table is directly reachable from a set of
tables, even though that set of tables does not include all the
dependencies of the lateral derived table.

The result is that DisallowParameterizedJoinPath() becomes too
aggressive and disallows too many join paths. In some cases, this
leads to indexes not being used in tables that are joined with lateral
tables, which hurts performance. In other cases, it disallows all
possible join paths, and the hypergraph optimizer is not able to come
up with any plan for the query.

This is fixed by making FindReachableTablesFrom() take the lateral
dependencies into account when checking if a lateral derived table is
reachable from a set of tables.

"SELECT COUNT(*) FROM information_schema.key_column_usage" now uses
indexes more efficiently when running with the hypergraph optimizer,
and its performance is on par with the old optimizer.

Change-Id: If3dce4800210ef84446a64f23c099144620be07e