WL#14843: Simplified hypergraph optimizer for simple queries [3/6, make hypergraph]
Most of the logic in MakeJoinHypergraph() is for handling joins.
Create a fast path through it that is used for single-table queries,
in which case there is no join to create a hypergraph for. The fast
path skips creating a join tree, extracting join predicates and
pushing predicates further down the join tree, as these steps are
completely unnecessary when there is no join.

Microbenchmark results (AMD EPYC 2.25GHz):

GCC 11.2.1:
BM_FindBestQueryPlanPointSelect       2641 -> 2348 ns/iter [+12.5%]

Clang 14.0.0:
BM_FindBestQueryPlanPointSelect       2827 -> 2513 ns/iter [+12.5%]

Change-Id: I164f6da84a6c4de995c6f8e24925fd4007ac2b38