WL#14843: Simplified hypergraph optimizer for simple queries [1/6, eq_ref]
When evaluating which access method to use for accessing a table, and
the query has predicates that compare each column of a unique index on
that table for equality with a constant, propose an eq_ref directly.
In such cases, it is unlikely that using any other access method will
give noticeably better performance, so planning time can be reduced by
only considering a single access method.

If there are multiple unique indexes that satisfy the criteria, an
eq_ref is proposed for each of them, and the tournament in
ProposeAccessPath() will decide which to keep based on the estimated
costs.

Time is saved primarily because the range optimizer is not invoked,
but also because we skip evaluating the cost of index scans and table
scans on the same table.

To avoid duplication of code in the new
ProposeAllUniqueIndexLookupsWithConstantKey() function, some checks
for applicability of a sargable predicate are moved from
ProposeRefAccess() to PossiblyAddSargableCondition(), so that
inapplicable predicates are not added to the list of sargable
predicates in the first place.

Microbenchmark results (AMD EPYC 2.25GHz):

GCC 11.2.1:
BM_FindBestQueryPlanPointSelect       4410 -> 2708 ns/iter [+63%]

Clang 14.0.0:
BM_FindBestQueryPlanPointSelect       4596 -> 2891 ns/iter [+59%]

Change-Id: I8dc01185de357ad94bf64ffb4d9ca25ff58abba9