Bug#34025164 AUTOTEST: testDict -n Bug13416603 I2 FAILING IN MULTIPLE CONFIGURATIONS
Problem:
--------

The failure is due to the presence of more index stat events
than the test expects. The number of events in the queue is
non-deterministic due to automatic stats updates which could
create additional events depending on when updates are
triggered in the background.

Fix:
----

Relax the checks related to number of events and clear the
event queue at the end of each case. This doesn't affect
the main focus of the test which is:

1. Handle node failure during index stat
2. Index stat when not all nodes are up (Bug#13416603)

Change-Id: I859ade16dafba6e46d03311e5f755ff90c53ae08