Bug#34039241 routertest_integration_routing_reuse gets stuck randomly on pb2.
When the server side of the connection that the Router handles is terminated
it sometimes happens that the client end gets stuck on a read.

This patch adds a canceling of the client side of the connection when
the server side is closed similar way the server side was caneclled
in case of client side being closed.
It also adds similar logic to the x-protocol connection handling.

Change-Id: Ic554c7eb014ce30e66a7d0a2ca74dee0718fa7f8