Bug#34038287 routertest_integration_routing_reuse fails randomly on pb2 u20.04
The test is launching mysqld process that tries to access the error log file
inside /var/log directory. This sometimes fails with "access denied" error
causing the mysqld process to fail.

This patch uses --log-error parameter to make the mysqld process use
test-provided temp directory for the error logging.
It also makes changes in the test setup to dump the Router and Server
logs in case of any test failures (not only the suite teardown fatal failures
as it was till now).

Change-Id: I0fb2e518a3ff0864ce5790d09dd3fab43ecd0872