Bug#34060289 optimizer rewrites view and turns left joins into inner joins
Problem:
We skip removing of constant condition if it has an Item_view_ref with
a subquery during fix_fields of an Item_cond.

But the next item of the where clause can contain a constant (without
an item_view_ref) which will trigger a removal and sets all the table
bits in the not_null_table_cache(~0).

Later (in simplify_joins) when tables are being evaluated to convert a
left join to an inner join, the used_tables in the
not_null_table_cache will enable the conversion to an inner join and
hence a wrong result.

Fix:
If an item in a condition is removed because it evaluates to always
true or always false, make sure the not_null_table_cache is also
empty.

Change-Id: I04529ac9e0ed3ffcf5b6b5f372f30b6c564166f7