Bug#34095231: hypergraph: Optimize away join node when one of the children is always empty
If the range optimizer detects that the predicates on a table will
always evaluate to false, the hypergraph optimizer uses a ZERO_ROWS
access path for that table. But if the table is part of a join, it
could be the case that the entire join, or possibly the entire query,
could be optimized away. The hypergraph optimizer does not optimize
away those other parts of the query.

This patch makes FoundSubgraphPair() check if either of the sides of
the join is always empty, and uses this to decide if the join should
also be replaced with a ZERO_ROWS access path. If the left side of an
inner join, a left join, a semijoin or an antijoin is always empty,
the join is also always empty. Similarly, if the right side of an
inner join or a semijoin is always empty, the join is always empty.

It also skips adding paths for grouping, ordering, windowing and
offset/limit if the join is always empty.

Change-Id: I84b9957a0b164f4def2b828ca26a03c7cc5a9dab