Bug #34050073 wl12527: write heavy server cores with meb when redo_log_capacity altered/ 60sec
InnoDB asserts that the combined size of all currently used redo log
files together with the currently unused ("temporary") files does not
exceed the current_physical_capacity, which is the last successfully set
redo log capacity.
However we have observed Assertion failure:
log0files_governor.cc:926:current_size <= current_capacity

This was puzzling, as a resize is not considered successful until each of
the files in use are smaller than current_physical_capacity/32, and the
combined size doesn't exceed current_physical_capacity, and when we
create a temporary file we ensure the new file has size
target_physical_capacity/32 and that all the files together will not
exceeed the current_physical_capacity total limit. So, how come that a
sum of sizes of files not larger than current_physical_capacity/32
exceeds the total?

Usually, InnoDB avoids having more than 32 files, but there's an
optimization which allows creation of a temporary, 33rd, redo log file
when there are exactly 32 files, the write position is "near" the end of
the 32nd file, and creating the new file would not exceed
current_physical_capacity. The goal of this optimization was to ensure
a smooth transition between file boundaries without log writer having to
wait for a new file to be provided.

There's another optimization, that when we resize redo log, we also
resize any existing temporary files to match the new target/32.

Unfortunatelly, these optimizations taken together violate the contract.
For example, suppose that initially, the situation is like this:
1 x 320kb  + 31 x 1280kb
and current_physical_capacity = 32x1280kb.
Now, suppose we temporarly resize redo log down back to 32x320kb, and that
the write position is near the end of the last big file.
Then we might create a temporary file, of size 1/32 of target capacity, so
it would have 320kb.
1 x 320kb  + 31 x 1280kb + 320kb (tmp)
This is allowed as it doesn't exceed the current capacity, which is still
roughly 32 x 1280kb (current_physical_capacity does not get smaller
immediatelly as we resize, only once we're done resizing, and we are not yet
done!).
Now, if we either cancel the resize or resize back to 32x1280kb, the
temporary file will get resized to 1280kb:
1 x 320kb  + 31 x 1280kb + 1280kb (tmp)
which is more than the current physical capacity, 32x1280kb, so we crash.

In this patch, we remove the second optimization.
There was already a piece of code which handled failures to resize the
temporary files by simply removing them. We will simply use it always.

Change-Id: Ic53890693b07b6de313c5a01f3a0e26723260beb