Bug #34051681 mysqld asan crash at ExplainMaterializeAccessPath
A query may have a predicate with a subquery. If that subquery
is never evaluated, no iterator objects will be created for the
Access_path objects that constitutes the plan of the subquery.

This lead to a segment violation when we tried to access one of these
non-existent iterators during 'EXPLAIN ANALYZE'.

This commit fixes that problem by checking if the iterator exists before
accessing it.

In addition, it ensures that we print "(never executed)" in the
"EXPLAIN ANALYZE" output for Access_path objects without an iterator.

Change-Id: Ic205514c22f5c8fcf39d70657e5e65e3f985edf5