Bug#33700735 - Server crashes and asserts in temptable code
Description
===========
The plan generated for a SELECT query with CTE involves
table materialization and index scan on the materialized
table. However, the temptable engine does not yet support
few index scan methods (index_first and index_last).

In case of other engines, the materialize access path has
special handling when the access path is not considered as
basic. INDEX_SCAN was not considered as basic and was trying
to find underlying basic access path and hence the undefined
behaviour.

Fix
===
Avoid index scan access methods on the temptable engine
tables. Consider the INDEX_SCAN access path as basic access
path.

Change-Id: Ie9330b18ad2f56b11827885872c8f0ca637d9abc