Bug#33992095: Add a microbenchmark for planning point selects in the hypergraph optimizer [fake handler]
Post-push fix. The Mock_HANDLER object in Fake_TABLE instruments the
calls to the handler functions, and this instrumentation disturbs the
microbenchmark. Add a fake handler instead, which mostly stubs out the
pure virtual functions and does not instrument the calls. This reduces
noise in the benchmark and makes it show more clearly the effect of
changing code in the hypergraph optimizer.

The microbenchmark runs ~40% faster with this change, so relatively
more of the time is spent in FindBestQueryPlan(), which is what it's
supposed to test.

Change-Id: Ib0e9457af093846d9b88c2b5fb4540787e48ecb7