Bug#33629814: Hypergraph Rapid side Crash : Incorrect Order By List
The interesting orders machinery in the hypergraph optimizer may
sometimes be able to elide the final sort for ORDER BY completely. But
in those cases where it is not able to eliminate the sort completely,
it creates a sort key based on the whole ORDER BY clause, including
any redundant elements that are there. This means the sort key could
contain elements that are redundant because of functional dependencies
on other elements of the sort key, and also completely meaningless
elements like constants.

This patch removes the redundant elements from the sort key before it
creates the sort node for the ORDER BY clause.

Change-Id: Iaa7a9cb9021778c4b1748cc380f12a92a2ce55aa