Bug#34092926 Emergency Failover after Network Split Invalidates Primary in ClusterSet
Router updates the metadata routers.last_check_in field periodically (every
10th metadata refresh)  with a current timestamp. This was added as a way of
helping users see which routers instances are in use and running.

The problem is that those updates on the metadata schema become errant
transactions whenever a split-brain occurs and make it impossible to rejoin
the old primary back to the ClusterSet. If there is a network split, Router
instances in a partition cannot understand that a failover happened so the
updates keep happening.

This patch changes the Router behavior so that the routers.last_check_in
fields gets  Update only once, when the Router starts. There are no periodic
updates done later. So the meaning of the last_check_in value is changing,
it should not anymore be read as the last time the Router was seen accessing
the metadata, it is the last time the Router was launched instead now.

Change-Id: Ided7b0f11092adf2635efe6b22fd087422aba0d4