Bug #34123159 Assertion failure: lock0lock.cc:5161:lock_rec_has_expl(LOCK_X | LOCK_REC_NOT_GAP
The fix for Bug #21356 "Change lock priority so that the transaction
holding S-lock gets X-lock first" made it possible for a transaction
requesting LOCK_X to bypass an already waiting LOCK_X request of another
transaction provided the later was already blocked by some other granted
lock of it.

There's also a very old mechanism which allows a transaction to elide
creating an explicit lock on a record if it's know it will obtain an
implicit lock on it - lock_rec_lock(true,...) sets impl=true which
refrains from creating a lock if there are no blockers. We use this
mechanism for example when an INSERT statement delete-unmarks an
existing record, trusting that setting TRX_ID header of the record is
enough to ensure implicit X_LOCK on it.

Taken together, a following scenario become possible:
1. t1 locks the record FOR SHARE,
2. t2 wants to lock it FOR UPDATE, so has to wait,
3. t1 wants to UPDATE it, and bypass t2 to avoid deadlock, and elides
   creation of an explicit lock as it expects to own implicit lock.

This leads to a state of the lock queue in which t2 has a
WAITING lock and t1 has an implicit lock, and the implicit lock is not
converted to explicit LOCK_X | LOCK_REC_NOT_GAP.
This state violates a very old debug-only validation assert, that if
impl_trx holds an implicit lock on a record, and there is any other lock
request for a lock on this record then impl_trx should also have an
explicit granted LOCK_X | LOCK_REC_NOT_GAP lock (or stronger).

It's not clear why it was so important to me despite reading comments,
logs, bug reports and code reviews around it - the oldest available
version of the code doesn't provide any rationale.

My educated guess is that the assert ensures that impl_trx owns not just
an implicit lock, but also a granted lock which guards against granting
the waiting lock too early. Also, it guards against the opposite case in
which the waiting lock is never granted as there is no explicit lock
release of which could cause the grant.

One approach would be to soften the assert to ensure only the two goals
mentioned in previous paragraph. The new assert would not care about the
exact lock mode of the impl_trx's lock, as long as it is granted and
blocks the other lock from being granted.
However this would require quite complicated proof of correctness:
we need to show that the explicit lock of impl_trx has a life cycle
which is at least as long as the one of the implicit X lock - which it
indeed has, because implicit locks are released earlier than explicit
locks during commit sequence. One would also have to show that an
explicit X lock is not needed to prevent other transactions from
acquiring an S lock - which is true, because if such S lock existed
before acquisition of implicit X lock, it would prevent the bypassing
heuristic from being used, and if S lock was created later then it
should cause implicit-to-explicit conversion of existing locks first.

Instead we choose a solution simpler to explain: we just create the
explicit X lock even if impl=true was passed to lock_rec_lock(..) in
cases where the bypassing heuristic had to be used.
This slightly changes the semantic of the impl argument, but apparently
the callers did not rely on the lock *not* being created.

Change-Id: Ic939fd205f7036bb4e026385cefdb8cf40cddb7f