Bug#34098384 Dbacc lock handling has too much overhead when WITH_DEBUG compiled
WITH_DEBUG compiled code had some self check code which required the
lockOwnersList to be fully iterated. For large transactions, holding many
locks, the overhead of this becomes prohobitive. The value of these self
check loops were considered to be questionable.

This patch replace the lockOwnersList with a simple lockCount which enables
us to verify that all locks had been released before removing a fragment.
Other self-check code checking the lockOwnersList for lock not already
held or release is just removed.

Related methods for inserting / removing a lock into the lockOwnersList
are removed as well.

Change-Id: I3182c154f1afb5d1bac847432d3ec1d1b5c0d3ba