Bug#32678303: Assertion 'table_num_to_node_num[table_num] != -1 failed Bug#34131822: Outer references in INSERT ... VALUES are not recognized
These bugs are fixed together because while fixing bug#32678303,
regressions in existing tests were seen due to bug#34131822,
which was then reported.

Bug#32678303 causes an outer reference in a derived table that points
to a table in a JOIN operation that is not available for the subquery
to be reported as valid, whereas a reference without a derived table
is correctly reported as invalid.

The reason for this is within PT_derived_table::contextualize(), where
the name resolution context for a derived table is erroneously set to
be the default context of the outer query block, rather than the
current context (which correctly identifies the available tables if
the derived table is placed inside a JOIN operation).

Fixed by setting up the correct name resolution context.

However, this problem was revealed after a mistake in the bug
simplification that found a similar problem for derived tables.
The original bug remained and was about a resolver problem for a CTE,
which required a separate fix of the name resolution context in
Query_block::find_common_table_expr().

Bug#34131822 means that an outer reference in a derived table placed
inside the VALUES clause of an INSERT statement is not recognized as
valid, whereas a similar reference without a derived table is
recognized. The problem is due to LEX::new_query(), where the outer
context was disregarded when used inside INSERT ... VALUES and
an INSERT ... ON DUPLICATE KEY UPDATE clause, however the first of
these should have a proper outer context.

Fixed by not making an exception for INSERT ... VALUES.

Also enabled a test in main.subquery_xx that was previously marked as
invalid, and removing an invalid comment in
i_main.window_functions-bug31389573.

Change-Id: Ic3a7c5252e6b8153a601eab70d3f0e089a9f6d30