BUG#34140697: Fix old terminology in help texts for mysqlcheck and mysqlupgrade
Replace old terminology by new terminology in help texts for
mysqlcheck and mysqlupgrade.

mysqlcheck: Just remove the mention of replicas. The use of binary
logs is broader than that.

mysqlupgrade: Remove the mention of replicas. And discourage the
option completely: servers are upgraded one at a time, replicas first,
so this only creates errant transactions. It does not "help downstream
replicas to prepare for upgrade" or anything.  It just risks breaking
replication after a fail-over.

Change-Id: I6d077a66ce4d92558a5e24d4b7deb72af64b6bc5