Bug#33899767 dict0load.cc:875 argument 2 null where non-null expected [-Werror=nonnull]
This error was reported while building mysql-trunk with valgrind
enabled. The cause of the error was null parameter being passed
to memcmp(). The parameter "last_index_id" passed to the function
"dict_load_field_low()" was literal nullptr.
Now the question is - Why did such a call exist in the first place?
Lets see the history of this function.

The function "dict_load_field_low()" used to have two calls:
one from "dict_process_sys_fields_rec()" and
another one from "dict_load_fields()" until 8.0.2.
In 8.0.3 we added the support for Atomic DDL and removed
old InnoDB system tables. The call with non-null "last_index_id"
was removed and it left just one call with null parameter. But
the original function was not refactored in this exercise.
This fix refactors the function "dict_load_field_low()"
and cleans up unused parameters and the associated code.

Change-Id: I7ca07ed03244754b8a7b5bcce30cb5b74adc6b46