Bug#34058984 during upgrade from MySQL Cluster 7.6 to 8.0.28 all triggers are gone
Problem:
--------

NDB tables are skipped in the MySQL Server upgrade phase and are
instead migrated by the ndbcluster plugin at a later stage. As a
result of this, triggers associated with NDB tables are not
created during upgrades from 5.7 based versions.

It isn't possible to create such triggers when the NDB tables
are migrated by the ndbcluster plugin as metadata about the
triggers is lost in the 'finalize upgrade' phase of the Server
upgrade where all .TRG files are deleted.

Fix:
----

1. Migration of NDB tables with triggers is not deferred during
   the Server upgrade phase.
2. NDB tables with triggers are not removed from the DD during
   setup even when initial system starts are detected.

Change-Id: Ic651545196f7228eca442e4acbcb30a185e6e903