Bug#33379204: Add a secondary engine flag to not rewrite field               references to temporary tables
Post-push fix for a problem with EXPLAIN on queries that use window
functions when the USE_EXTERNAL_EXECUTOR secondary engine flag is
enabled.

EXPLAIN gets information about the window from the temporary table
parameters in the WINDOW access path. When the USE_EXTERNAL_EXECUTOR
flag is enabled, the temporary table is not created, so EXPLAIN would
crash.

Fixed by using the "window" member of the WINDOW access path instead
of "temp_table_param" when explaining the query, and making the old
optimizer initialize this member of the access path (the hypergraph
optimizer already did).

Also changed the original test case for this bug to inspect the
returned access path more closely. (Not directly related to the rest
of this patch.)

Change-Id: I55f787e0de9b6ba2a4b6b70fd1b64534ab3421b0